MiyanKey/requirements.txt
pyTelegramBotAPI
Flask
python-dotenv
pymongo
schedule